Task #216344 - UI for registration page
