+++ package.json
@@ -27,10 +27,13 @@
   "dependencies": {
     "@chakra-ui/icons": "^2.1.1",
     "@chakra-ui/react": "^2.8.0",
-    "@emotion/react": "^11.11.1",
+    "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
     "@hookform/resolvers": "^3.3.1",
-    "@mui/material": "^5.14.7",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/joy": "^5.0.0-beta.32",
+    "@mui/material": "^5.15.14",
     "ag-grid-community": "^30.1.0",
     "antd": "^5.9.3",
     "axios": "^1.6.7",
